requirements.txt
anthropic>=0.40.0
rich>=13.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
python-dotenv>=1.0.0
streamlit>=1.32.0
plotly>=5.18.0
PyJWT>=2.8.0
cryptography>=41.0.0
